requirements.txt
python-dotenv==0.16.0
requests==2.27.1
play-sounds==0.5.0
